# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.AreaOfInterest:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      events:
        items:
          $ref: '#/definitions/models.Event'
        type: array
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      polygon_area:
        type: string
      radius_in_meters:
        type: number
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/models.User'
      created_by_id:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      event_id:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Event:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by_id:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      device:
        $ref: '#/definitions/models.GPSDevice'
      device_id:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      media_presigned_urls:
        items:
          $ref: '#/definitions/models.PresignedUrl'
        type: array
      status:
        $ref: '#/definitions/models.EventStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/models.EventType'
      updated_at:
        type: string
    type: object
  models.EventStatus:
    enum:
    - open
    - resolved
    - closed
    type: string
    x-enum-varnames:
    - EventStatusOpen
    - EventStatusResolved
    - EventStatusClosed
  models.EventType:
    enum:
    - robbery
    - lost
    - accident
    - other
    type: string
    x-enum-varnames:
    - EventTypeRobbery
    - EventTypeLost
    - EventTypeAccident
    - EventTypeOther
  models.GPSDevice:
    properties:
      api_cookie:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: string
      imei:
        type: string
      locations:
        items:
          $ref: '#/definitions/models.GPSLocation'
        type: array
      name:
        type: string
      number:
        type: string
      password:
        type: string
      tracking:
        type: boolean
      updated_at:
        type: string
    type: object
  models.GPSLocation:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      device_id:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      updated_at:
        type: string
    type: object
  models.MediaFile:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      key:
        type: string
      updated_at:
        type: string
    type: object
...
  models.PresignedUrl:
    properties:
      key:
        type: string
      url:
        type: string
    type: object
  models.User:
    properties:
      areas_of_interest:
        items:
          $ref: '#/definitions/models.AreaOfInterest'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  schemas.CreateAreaOfInterest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      polygon_area:
        type: string
      radius_in_meters:
        type: number
    type: object
  schemas.CreateComment:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
  schemas.CreateEvent:
    properties:
      communities:
        items:
          type: string
        type: array
      description:
        type: string
      device_id:
        type: string
      is_public:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      status:
        $ref: '#/definitions/models.EventStatus'
      title:
        type: string
      type:
        $ref: '#/definitions/models.EventType'
    required:
    - description
    - title
    type: object
...
  schemas.Error:
    properties:
      error:
        type: string
    type: object
...
  schemas.UpdateComment:
    properties:
      content:
        type: string
    type: object
...
  schemas.UpdateEvent:
    properties:
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      status:
        $ref: '#/definitions/models.EventStatus'
      title:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/comments/{comment_id}:
    delete:
      description: Delete a comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Delete a comment
      tags:
      - comments
    patch:
      consumes:
      - application/json
      description: Update a comment by its ID
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: Update comment
        in: body
        name: updateComment
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateComment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Update a comment
      tags:
      - comments
...
  /api/events:
    post:
      consumes:
      - application/json
      description: Create a new event for the currently authenticated user
      parameters:
      - description: Create event
        in: body
        name: createEvent
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateEvent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Create a new event
      tags:
      - events
  /api/events/{id}:
    delete:
      description: Delete an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Delete an event
      tags:
      - events
    get:
      description: Get details of an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Get an event by ID
      tags:
      - events
    patch:
      consumes:
      - application/json
      description: Update an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Update event
        in: body
        name: updateEvent
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateEvent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Update an event
      tags:
      - events
  /api/events/{id}/comments:
    get:
      description: Get all comments for an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Get comments for an event
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Post a comment on an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Create comment
        in: body
        name: createComment
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateComment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Post a comment on an event
      tags:
      - comments
  /api/events/{id}/media:
    post:
      consumes:
      - multipart/form-data
      description: Upload media to an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MediaFile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/schemas.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Upload media to an event
      tags:
      - events
  /api/events/from-area:
    post:
      description: Get events in an area by its ID
      parameters:
      - description: Create area of interest
        in: body
        name: createAreaOfInterest
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateAreaOfInterest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Error'
      summary: Get events in an area
      tags:
      - events
...
swagger: "2.0"